requirements.txt
groq==0.9.0
streamlit>=1.35.0
python-dotenv>=1.0.0
soundfile>=0.12.1